% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Add_CostBenefit_Axis.R
\name{Add_CostBenefit_Axis}
\alias{Add_CostBenefit_Axis}
\title{Add cost benefit ratio axis to a decision curve plot.}
\usage{
Add_CostBenefit_Axis(xlim, cost.benefits, n.cost.benefits = 6, line = 4,
  ...)
}
\arguments{
\item{xlim}{}

\item{cost.benefits}{Character vector of the form c("c1:b1", "c2:b2", ..., "cn:bn") with integers ci, bi corresponding to specific cost:benefit ratios to print.}

\item{n.cost.benefits}{number of cost:benefit ratios to print if cost.benefit.axis = TRUE (default n.cost.benefit = 6).}

\item{line}{x-axis line to print the axis (default is 4).}
}
\value{
List with components
}
\description{
Add cost benefit ratio axis to a decision curve plot.
}
\examples{
#helper function
expit <- function(xx) exp(xx)/ (1+exp(xx))

#load simulated data
data(dcaData)
# Assume we have access to previously published models
# (or models built using a training set)
# that we can use to predict the risk of cancer.

# Basic model using demographic variables: Age, Female, Smokes.
dcaData$BasicModel <- with(dcaData, expit(-7.3 + 0.18*Age - 0.08*Female + 0.80*Smokes ) )

# Model using demographic + markers : Age, Female, Smokes, Marker1 and Marker2.
dcaData$FullModel <- with(dcaData, expit(-10.5 + 0.22*Age  - 0.01*Female + 0.91*Smokes + 2.03*Marker1 - 1.56*Marker2))

#use DecisionCurve defaults (set bootstraps = 25 here to reduce computation time).
DecisionCurve(dcaData,
            outcome = "Cancer", predictors = c("BasicModel", "FullModel"),
             bootstraps = 25)

par(mar = c(7.5, 4, 3, 2) + 0.1)
(finish later)
}
\seealso{
\code{\link{summary.DecisionCurve}},  \code{\link{DecisionCurve}}
}

