% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_decision_curve.R
\name{cv_decision_curve}
\alias{cv_decision_curve}
\title{Calculate cross-validated decision curves}
\usage{
cv_decision_curve(formula, data, family = binomial(link = "logit"),
  thresholds = seq(0, 1, by = 0.01), folds = 5, study.design = c("cohort",
  "case-control"), population.prevalence)
}
\arguments{
\item{formula}{an object of class 'formula' of the form outcome ~ predictors, giving the prediction model to be fitted using glm.}

\item{data}{data.frame containing outcome and predictors. Missing data on any of the predictors will cause the entire observation to be removed.}

\item{family}{a description of the error distribution and link function to pass to 'glm' used for model fitting. Defaults to binomial(link = "logit") for logistic regression.}

\item{thresholds}{Numeric vector of high risk thresholds to use when plotting and calculating net benefit values.}

\item{folds}{Number of folds for k-fold cross-validation.}

\item{study.design}{Either 'cohort' (default) or 'case-control' describing the study design used to obtain data. See details for more information.}

\item{population.prevalence}{Outcome prevalence rate in the population used to calculate decision curves when study.design = 'case-control'.}
}
\value{
List with components
\itemize{
  \item derived.data: A data frame in long form showing the following for each predictor and each 'threshold', 'FPF':false positive fraction, 'TPF': true positive fraction, 'NB': net benefit, 'sNB': standardized net benefit, 'rho': outcome prevalence, 'predictor': name of predictor, 'xx_lower', 'xx_upper': the lower and upper confidence bands for TPF, FPF, rho, NB and sNB.
  \item standardized: Whether standardized net benefit or net benefit is returned.
  \item call: matched function call.
}
}
\description{
This is a wrapper for 'decision_curve' that computes k-fold cross-validated estimates of sensitivity, specificity, and net benefit so that cross-validated net benefit curves can be plotted.
}
\examples{

#todo

}
\seealso{
\code{\link{summary.decision_curve}}, \code{\link{decision_curve}},  \code{\link{Add_CostBenefit_Axis}}
}

