% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dcaData}
\alias{dcaData}
\title{Simulated dataset for package 'DecisionCurve'}
\format{A data frame with 500 rows and 6 variables:
\itemize{
  \item Age: Age in years.
  \item Female: Indicator for female gender.
  \item Smokes: Indicator for smoking status.
  \item Marker1: simulated biomarker.
  \item Marker2: simulated biomarker.
  \item Cancer: Indicator for cancer.
}}
\usage{
dcaData
}
\description{
Simulated cohort data containing demographic variables,
marker values and cancer outcome.
}
\keyword{datasets}
