% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dcaData_cc}
\alias{dcaData_cc}
\title{Simulated dataset for package 'DecisionCurve'}
\format{A data frame with 500 rows and 6 variables:
\itemize{
  \item Age: Age in years.
  \item Female: Indicator for female gender.
  \item Smokes: Indicator for smoking status.
  \item Marker1: simulated biomarker.
  \item Marker2: simulated biomarker.
  \item Cancer: Indicator for cancer.
}}
\usage{
dcaData_cc
}
\description{
Simulated data from a case-control study containing demographic variables,
marker values and cancer outcome. The prevalence of cancer
in the population that this data is sampled from is approximately 0.11.
}
\keyword{datasets}
