% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DecisionCurve.R
\name{DecisionCurve}
\alias{DecisionCurve}
\title{Plot decision curves}
\usage{
DecisionCurve(data, outcome, predictors, thresholds = seq(0, 1, by = 0.01),
  standardize = TRUE, confidence.intervals = 0.95, bootstraps = 500,
  cost.benefit.axis = TRUE, n.cost.benefits = 6, cost.benefits,
  legend.position = "topright", col, lty, lwd = 2, xlim, ylim,
  xlab = "Risk Threshold", ylab, cost.benefit.xlab = "Cost:Benefit Ratio",
  plot = TRUE, ...)
}
\arguments{
\item{data}{data.frame containing outcome and predictors. Missing data on any of the predictors will cause the entire observation to be removed.}

\item{outcome}{Name of outcome of interest found in data. Within 'data', the outcome variable must be a numeric vector of 0's and 1's.}

\item{predictors}{Vector of names for risk predictors to calculate decision curves. 'data' must have columns corresponding to the names provided consisting of predicted risks Pr(outcome = 1 | predictor). Risks must fall between 0 and 1.}

\item{thresholds}{Numeric vector of high risk thresholds to use when plotting and calculating net benefit values.}

\item{standardize}{logical (default TRUE) indicating whether to use the standardized net benefit (NB/disease prevalence) or not.}

\item{confidence.intervals}{Numeric (default 0.95 for 95\% confidence bands) level of bootstrap confidence intervals to plot. Set as NA or 'none' to remove confidence intervals. See details for more information.}

\item{bootstraps}{Number of bootstrap replicates to use to calculate confidence intervals (default 500).}

\item{cost.benefit.axis}{logical (default TRUE) indicating whether to print an additional x-axis showing relative cost:benefit ratios in addition to risk thresholds.}

\item{n.cost.benefits}{number of cost:benefit ratios to print if cost.benefit.axis = TRUE (default n.cost.benefit = 6).}

\item{cost.benefits}{Character vector of the form c("c1:b1", "c2:b2", ..., "cn:bn") with integers ci, bi corresponding to specific cost:benefit ratios to print.}

\item{legend.position}{One of "topright" (default), "top", "topleft", "left", "bottomleft", "bottom", "bottomright", "right", or "none" indicating where the legend should be printed.}

\item{col}{vector of color names to be used in plotting corresponding to the 'predictors' given. Default colors will be chosen from rainbow(..., v = .8). See details for more information on plot parameters.}

\item{lty}{vector of linetypes to plot corresponding to 'predictors'.}

\item{lwd}{vector of linewidths to plot corresponding to 'predictors'.}

\item{xlim}{vector giving c(min, max) of x-axis. Defaults to c(min(thresholds), max(thresholds)).}

\item{ylim}{vector giving c(min, max) of y-axis.}

\item{xlab}{label of main x-axis.}

\item{ylab}{label of y-axis.}

\item{cost.benefit.xlab}{label of cost:benefit ratio axis.}

\item{plot}{Logical (default TRUE) indicating whether to plot decision curves.}

\item{...}{other options directly send to plot()}
}
\value{
List with components
\itemize{
  \item plot.data: a list with further data.frame components 'estimates', 'ci_lower', and 'ci_upper' showing point estimates of (standardized) net benefits, and lower and upper ci's, respectively.
  \item derived.data: A data frame in long form showing the following for each predictor and each 'threshold', 'FPF':false positive fraction, 'TPF': true positive fraction, 'NB': net benefit, 'sNB': standardized net benefit, 'rho': outcome prevalence, 'predictor': name of predictor, 'x_lower', 'x_upper': the lower and upper confidence bands for NB or sNB.
  \item standardized: Whether standardized net benefit or net benefit is returned.
  \item call: matched function call.
}
}
\description{
Decision curves are a useful tool to evaluate the population impact of adopting a risk prediction instrument into clinical practice. Given one or more instruments (risk models) that predict the probability of a binary outcome, this package calculates and plots decision curves, which display estimates of the standardized net benefit by the probabilty threshold used to categorize observations as 'high risk.'  Bootstrap confidence intervals are displayed as well. This package is a companion to the manuscript 'Kerr. et. al (put ref here.)'.
}
\details{
Confidence intervals for (standardized) net benefit are calculated pointwise at each risk threshold. Bootstrap sampling is done without stratifying on outcome, so disease prevalence varies within bootstrap samples. Note that the confidence intervals calculated assume the risk prediction tool
}
\examples{
#helper function
expit <- function(xx) exp(xx)/ (1+exp(xx))

#load simulated data
data(dcaData)
# Assume we have access to previously published models
# (or models built using a training set)
# that we can use to predict the risk of cancer.

# Basic model using demographic variables: Age, Female, Smokes.
dcaData$BasicModel <- with(dcaData, expit(-7.3 + 0.18*Age - 0.08*Female + 0.80*Smokes ) )

# Model using demographic + markers : Age, Female, Smokes, Marker1 and Marker2.
dcaData$FullModel <- with(dcaData, expit(-10.5 + 0.22*Age  - 0.01*Female + 0.91*Smokes + 2.03*Marker1 - 1.56*Marker2))

#use DecisionCurve defaults (set bootstraps = 25 here to reduce computation time).
DecisionCurve(dcaData,
            outcome = "Cancer", predictors = c("BasicModel", "FullModel"),
             bootstraps = 25)
}
\seealso{
\code{\link{summary.DecisionCurve}},  \code{\link{Add_CostBenefit_Axis}}
}

