% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions_main.R
\name{plot_decision_curve}
\alias{plot_decision_curve}
\title{Plot a decision.curve object or many decision.curve objects}
\usage{
plot_decision_curve(x, curve.names, cost.benefit.axis = TRUE,
  n.cost.benefits = 6, cost.benefits, standardize = TRUE,
  confidence.intervals, col, lty, lwd = 2, xlim, ylim,
  xlab = "Risk Threshold", ylab, cost.benefit.xlab = "Cost:Benefit Ratio",
  legend.position = c("topright", "right", "bottomright", "bottom",
  "bottomleft", "left", "topleft", "top", "none"), ...)
}
\arguments{
\item{x}{DecisionCurve object to plot or a list of DecisionCurve objects. Assumes output from function 'DecisionCurve'}

\item{cost.benefit.axis}{logical (default TRUE) indicating whether to print an additional x-axis showing relative cost:benefit ratios in addition to risk thresholds.}

\item{n.cost.benefits}{number of cost:benefit ratios to print if cost.benefit.axis = TRUE (default n.cost.benefit = 6).}

\item{cost.benefits}{Character vector of the form c("c1:b1", "c2:b2", ..., "cn:bn") with integers ci, bi corresponding to specific cost:benefit ratios to print.}

\item{standardize}{logical (default TRUE) indicating whether to use the standardized net benefit (NB/disease prevalence) or not.}

\item{col}{vector of color names to be used in plotting corresponding to the 'predictors' given. Default colors will be chosen from rainbow(..., v = .8). See details for more information on plot parameters.}

\item{lty}{vector of linetypes to plot corresponding to 'predictors'.}

\item{lwd}{vector of linewidths to plot corresponding to 'predictors'.}

\item{xlim}{vector giving c(min, max) of x-axis. Defaults to c(min(thresholds), max(thresholds)).}

\item{ylim}{vector giving c(min, max) of y-axis.}

\item{xlab}{label of main x-axis.}

\item{ylab}{label of y-axis.}

\item{cost.benefit.xlab}{label of cost:benefit ratio axis.}

\item{...}{other arguments ignored (for compatibility with generic)}

\item{nround}{number of decimal places to round (default 3).}

\item{...}{other options directly send to plot()}
}
\description{
Plot a decision.curve object or many decision.curve objects
}
\examples{
#helper function
expit <- function(xx) exp(xx)/ (1+exp(xx))

#load simulated data
data(dcaData)
# Assume we have access to previously published models
# (or models built using a training set)
# that we can use to predict the risk of cancer.

# Basic model using demographic variables: Age, Female, Smokes.
dcaData$BasicModel <- with(dcaData, expit(-7.3 + 0.18*Age - 0.08*Female + 0.80*Smokes ) )

# Model using demographic + markers : Age, Female, Smokes, Marker1 and Marker2.
dcaData$FullModel <- with(dcaData, expit(-10.5 + 0.22*Age  - 0.01*Female + 0.91*Smokes + 2.03*Marker1 - 1.56*Marker2))

#use DecisionCurve defaults (set bootstraps = 25 here to reduce computation time).
my.dc <- DecisionCurve(dcaData,
                      outcome = "Cancer", predictors = c("BasicModel", "FullModel"),
                      bootstraps = 25,
                      thresholds = seq(0, 0.4, by = .05))

summary(my.dc)
}

