% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.DecisionCurve.R
\name{summary.DecisionCurve}
\alias{summary.DecisionCurve}
\title{Displays a useful description of a DecisionCurve object}
\usage{
\method{summary}{DecisionCurve}(x, ..., nround = 3)
}
\arguments{
\item{...}{other arguments ignored (for compatibility with generic)}

\item{nround}{number of decimal places to round (default 3).}

\item{object}{DecisionCurve object to summarise}
}
\description{
Displays a useful description of a DecisionCurve object
}
\examples{
#helper function
expit <- function(xx) exp(xx)/ (1+exp(xx))

#load simulated data
data(dcaData)
# Assume we have access to previously published models
# (or models built using a training set)
# that we can use to predict the risk of cancer.

# Basic model using demographic variables: Age, Female, Smokes.
dcaData$BasicModel <- with(dcaData, expit(-7.3 + 0.18*Age - 0.08*Female + 0.80*Smokes ) )

# Model using demographic + markers : Age, Female, Smokes, Marker1 and Marker2.
dcaData$FullModel <- with(dcaData, expit(-10.5 + 0.22*Age  - 0.01*Female + 0.91*Smokes + 2.03*Marker1 - 1.56*Marker2))

#use DecisionCurve defaults (set bootstraps = 25 here to reduce computation time).
my.dc <- DecisionCurve(dcaData,
                      outcome = "Cancer", predictors = c("BasicModel", "FullModel"),
                      bootstraps = 25,
                      thresholds = seq(0, 0.4, by = .05))

summary(my.dc)
}

